newsfetcher/requirements.txt
newsapi-python == 0.2.7
dotenv == 0.9.9
python-dotenv~=1.2.1
redis == 7.0.1